finance_advisor/requirements.txt
google-adk[extensions]
google-generativeai 
python-dotenv
litellm
pydantic 
requests
google-cloud-aiplatform[adk,agent_engines]